+++ R/mf_theme.R
@@ -2,6 +2,7 @@
 #' @description This function set a map theme.
 #' The parameters set by this function are the figure margins, background and
 #' foreground colors and some \link{mf_title} options.
+#' Use \code{mf_theme('default')} to reset to default theme settings.
 #' @param x name of a map theme. One of "default", "brutal", "ink",
 #' "dark", "agolalight", "candy", "darkula", "iceberg", "green", "nevermind",
 #' "jsk", "barcelona". If x is used other parameters are ignored.
@@ -17,7 +18,6 @@
 #' @details
 #' It is also possible to set a custom theme using a list of arguments
 #' (see Examples).
-#' Use \code{mf_theme('default')} to reset theme settings.
 #' \code{mf_theme()} returns the current theme settings.
 #' @return The (invisible) list of theme parameters is returned.
 #' @export
@@ -214,11 +214,12 @@ mf_theme <- function(x = "default", bg, fg, mar, tab, pos, inner, line, cex,
     theme <- x
   } else {
     if (!x %in% names(themes)) {
-      stop(paste0(
-        "x should be one of ",
-        paste0(names(themes), collapse = ", ")
-      ),
-      call. = FALSE
+      stop(
+        paste0(
+          "x should be one of ",
+          paste0(names(themes), collapse = ", ")
+        ),
+        call. = FALSE
       )
     } else {
       theme <- themes[[x]]